pyproject.toml
[tool.poetry]
name = "tea-agent"
version = "0.1.0"
description = "AI agent for real-time tea ceremony analysis using camera vision"
authors = ["Claude AI"]

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.4.0"
opencv-python = "^4.10.0"
numpy = "^2.1.0"
google-cloud-aiplatform = "^1.58.0" # vertexaiを含むパッケージ
google-generativeai = "^0.8.5"
google-adk = "^1.13.0"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"